requirements.txt
streamlit
pinecone-io
langchain
python-dotenv